user_auth/pyproject.toml
[tool.poetry]
name = "user-auth"
version = "0.1.0"
description = ""
authors = ["Muhammad Saad <saadjameel2001@gmail.com>"]
readme = "README.md"
package-mode =false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.20"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.10.4"}
uvicorn = "^0.34.0"
streamlit = "^1.41.1"
extra-streamlit-components = "^0.1.71"
bcrypt = "3.2.0"
pydantic-settings = "^2.7.1"
psycopg = "^3.2.4"
poetry-plugin-export = "^1.9.0"
fastapi-sqlalchemy = "^0.2.1"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
